refactor: user정보이동
# Conflicts:
#	src/main/java/com/followMe/delivery_server/delivery/presentation/ShipmentController.java
