+++ src/main/java/com/followMe/delivery_server/delivery/domain/UserContext.java
@@ -1,4 +1,4 @@
-package com.followMe.delivery_server.delivery.application;
+package com.followMe.delivery_server.delivery.domain;
 
 import java.util.UUID;
