+++ src/main/java/com/followMe/delivery_server/delivery/domain/UserRole.java
@@ -1,4 +1,4 @@
-package com.followMe.delivery_server.delivery.application;
+package com.followMe.delivery_server.delivery.domain;
 
 public enum UserRole {
   VENDOR,
